security rework - Refactored SecurityConfig and separated authentication to a new AuthConfig
Reorganized SecurityConfig by removing authentication configurations and moving them into a newly created AuthConfig file. This included password encoding and token verification. This change is aimed at improving code structure and making it more efficient and maintainable. Also added new fields and methods in SecurityConfig for improving context handling.
